+++ policyengine_uk/parameters/gov/dfe/study_childcare_entitlement/care_to_learn/amount.yaml
@@ -1,0 +1,24 @@
+description: The Department for Education provides financial support through the Care to Learn scheme.
+in_london:
+  description: The maximum weekly amount available per child for those living in London.
+  metadata:
+    unit: currency-GBP
+    period: week
+    label: maximum weekly amount per child in london
+    reference:
+      - title: Care to Learn scheme
+        href: https://www.gov.uk/care-to-learn
+  values:
+    2024-01-01: 195.0
+
+outside_london:
+  description: The maximum weekly amount available per child for those living outside London.
+  metadata:
+    unit: currency-GBP
+    period: week
+    label: maximum weekly amount per child outside london
+    reference:
+      - title: Care to Learn scheme
+        href: https://www.gov.uk/care-to-learn
+  values:
+    2024-01-01: 180.0
