--- policyengine_uk/parameters/gov/hmrc/study_childcare_entitlement/care_to_learn_age_eligible.yaml
@@ -1,10 +1,0 @@
-description: The Department for Education sets a maximum age for carers to be eligible for care to learn scheme.
-metadata:
-  unit: year
-  period: year
-  label: Maximum age to be eligible for care to learn scheme
-  reference:
-    - title: none
-      href: none
-values:
-  2015-01-01: 20
