--- policyengine_uk/parameters/gov/hmrc/study_childcare_entitlement/care_to_learn_amount.yaml
@@ -1,24 +1,0 @@
-description: Amount of support available through the Care to Learn scheme
-in_london:
-  description: Maximum weekly amount available per child for those living in London
-  metadata:
-    unit: currency-GBP
-    period: week
-    label: Maximum weekly amount per child in London
-    reference:
-      - title: none
-        href: none
-  values:
-    2024-01-01: 195.0
-
-outside_london:
-  description: Maximum weekly amount available per child for those living outside London
-  metadata:
-    unit: currency-GBP
-    period: week
-    label: Maximum weekly amount per child outside London
-    reference:
-      - title: none
-        href: none
-  values:
-    2024-01-01: 180.0
