+++ policyengine_uk/variables/gov/dfe/study_childcare_entitlement/care_to_learn_eligible.py
@@ -4,18 +4,18 @@ from policyengine_uk.model_api import *
 class care_to_learn_eligible(Variable):
     value_type = bool
     entity = Person
-    label = "Eligible for Care to Learn childcare support"
-    documentation = "Whether the person is eligible for Care to Learn support"
+    label = "eligible for care to learn childcare support"
+    documentation = "Whether the person is eligible for Care to Learn support."
     definition_period = YEAR
 
     def formula(person, period, parameters):
         # Only parents can be eligible, not children
-        is_parent = ~person("is_child", period)
+        is_child = person("is_child", period)
 
         # Check basic eligibility conditions
         has_children = person.benunit.any(person("is_child", period))
-        p = parameters(period).gov.hmrc.study_childcare_entitlement
-        under_20 = person("age", period) < p.care_to_learn_age_eligible
+        p = parameters(period).gov.dfe.study_childcare_entitlement.care_to_learn
+        age_eligible = person("age", period) < p.age_limit
 
         current_ed = person("current_education", period)
         education_types = current_ed.possible_values
@@ -32,9 +32,9 @@ class care_to_learn_eligible(Variable):
 
         # Return eligibility - must be a parent AND meet all other criteria
         return (
-            is_parent
+            ~is_child
             & has_children
-            & under_20
+            & age_eligible
             & not_higher_education
             & lives_in_england
             & not_apprentice
